Move global chart context hooks into separate files
