+++ projects/js-packages/charts/src/components/legend/use-chart-legend-data.ts
@@ -1,5 +1,5 @@
 import { useMemo } from 'react';
-import { useGlobalChartTheme } from '../../hooks';
+import { useGlobalChartTheme } from '../../providers/chart-context';
 import { getItemShapeStyles, getSeriesStroke } from '../../utils';
 import type { BaseLegendItem } from './types';
 import type { ChartTheme, SeriesData, DataPointDate, DataPointPercentage } from '../../types';
