+++ projects/js-packages/charts/src/hooks/index.ts
@@ -1,5 +1,4 @@
 export { useDeepMemo } from './use-deep-memo';
-export { useGlobalChartTheme } from './use-global-chart-theme';
 export { useChartMouseHandler } from './use-chart-mouse-handler';
 export { useXYChartTheme } from './use-xychart-theme';
 export { useChartDataTransform } from './use-chart-data-transform';
