+++ projects/js-packages/charts/src/index.ts
@@ -19,12 +19,13 @@ export { LineShape, CircleShape, RectShape } from './visx/legend';
 // Themes
 export { ThemeProvider } from './providers/theme';
 export { defaultTheme, jetpackTheme, wooTheme } from './providers/theme/themes';
+export { useGlobalChartTheme } from './providers/chart-context';
 
 // Utilities
 export { mergeThemes } from './utils';
 
 // Hooks
-export { useDeepMemo, useGlobalChartTheme, useChartMouseHandler, useXYChartTheme } from './hooks';
+export { useDeepMemo, useChartMouseHandler, useXYChartTheme } from './hooks';
 
 // LeaderboardChart utilities
 export { formatMetricValue } from './components/leaderboard-chart';
