+++ projects/js-packages/charts/src/providers/chart-context/hooks/use-chart-id.ts
@@ -1,0 +1,6 @@
+import { useId } from 'react';
+
+export const useChartId = ( providedId?: string ): string => {
+	const generatedId = useId();
+	return providedId || generatedId;
+};
