+++ projects/js-packages/charts/src/providers/chart-context/index.ts
@@ -1,7 +1,6 @@
-export {
-	GlobalChartsProvider,
-	GlobalChartsContext,
-	useGlobalChartsContext,
-} from './global-charts-provider';
-export { useChartId, useChartRegistration } from './utils';
+export { GlobalChartsProvider, GlobalChartsContext } from './global-charts-provider';
+export { useGlobalChartContext } from './hooks/use-global-chart-context';
+export { useChartId } from './hooks/use-chart-id';
+export { useChartRegistration } from './hooks/use-chart-registration';
+export { useGlobalChartTheme } from './hooks/use-global-chart-theme';
 export type { ChartContextValue, ChartRegistration } from './types';
